pyproject.toml
[tool.poetry]
name = "image_to_skill"
version = "0.1.0"
description = ""
authors = ["Woong Ze Yi <woongzeyi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.3.2"
filetype = "^1.0.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyinstaller = "^4.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
